keep 'Add files' and 'Add folder' controls together
